# Separate RestartDictionaries and RestartCodecs frame flags