--- a/go/otel/oteltef/eventarray.go
+++ b/go/otel/oteltef/eventarray.go
@@ -260,6 +260,7 @@ func (d *EventArrayDecoder) Continue() {
 }
 
 func (d *EventArrayDecoder) Reset() {
+	d.prevLen = 0
 	d.decoder.Reset()
 }