--- a/go/otel/oteltef/int64array.go
+++ b/go/otel/oteltef/int64array.go
@@ -260,6 +260,7 @@ func (d *Int64ArrayDecoder) Continue() {
 }
 
 func (d *Int64ArrayDecoder) Reset() {
+	d.prevLen = 0
 	d.decoder.Reset()
 }