--- a/go/otel/oteltef/linkarray.go
+++ b/go/otel/oteltef/linkarray.go
@@ -260,6 +260,7 @@ func (d *LinkArrayDecoder) Continue() {
 }
 
 func (d *LinkArrayDecoder) Reset() {
+	d.prevLen = 0
 	d.decoder.Reset()
 }