--- a/go/otel/oteltef/pointvalue.go
+++ b/go/otel/oteltef/pointvalue.go
@@ -293,8 +293,6 @@ func (e *PointValueEncoder) Init(state *WriterState, columns *pkg.WriteColumnSet
 }
 
 func (e *PointValueEncoder) Reset() {
-	// Since we are resetting the state of encoder make sure the next Encode()
-	// call forcedly writes all fields and does not attempt to skip.
 	e.prevType = 0
 	e.int64Encoder.Reset()
 	e.float64Encoder.Reset()