--- a/go/pkg/frameflags.go
+++ b/go/pkg/frameflags.go
@@ -7,6 +7,8 @@ const (
 	RestartDictionaries FrameFlags = 1 << iota
 	// RestartCompression resets and restarts the compression stream at frame beginning.
 	RestartCompression
+	// RestartCodecs resets the state of all encoders/decoders at frame beginning.
+	RestartCodecs
 
-	FrameFlagsMask = RestartDictionaries | RestartCompression
+	FrameFlagsMask = RestartDictionaries | RestartCompression | RestartCodecs
 )