--- a/stefgen/templates/reader.go.tmpl
+++ b/stefgen/templates/reader.go.tmpl
@@ -89,6 +89,12 @@ func (r *{{.StructName}}Reader) nextFrame() error {
 		r.state.ResetDicts()
 	}
 
+	if frameFlags&pkg.RestartCodecs != 0 {
+		// The frame that has just started indicates that the decoders
+		// must be restarted.
+		r.decoder.Reset()
+	}
+
 	r.decoder.Continue()
 	return nil
 }