Fixes https://github.com/splunk/stef/issues/47

Previously the flags were joint and incorrectly implemented. The writers previously always reset dictionaries and encoders at the same time, while readers only reset the dictionaries.

Now the flags are separate and each indicates the resetting of either dictionaries or codecs. The writers and readers correctly implement each flag.

Enabled previously disabled randomized write/read tests, which were failing because of incorrect implementation. Also added a few more randomized tests for other combinations of write options.

There is no performance impact for this change:
```
                                │     main     │                new                │
                                │    sec/op    │   sec/op     vs base              │
SerializeNative/STEF/none-10      3.952m ± 23%   4.030m ± 4%       ~ (p=0.485 n=6)
DeserializeNative/STEF/none-10    1.764m ±  1%   1.756m ± 1%       ~ (p=0.818 n=6)
SerializeFromPdata/STEF/none-10   90.15m ±  1%   91.12m ± 4%       ~ (p=0.132 n=6)
DeserializeToPdata/STEF/none-10   22.66m ± 17%   22.62m ± 2%       ~ (p=0.699 n=6)
geomean                           10.92m         10.99m       +0.60%

                                │     main     │                new                │
                                │  sec/point   │  sec/point   vs base              │
SerializeNative/STEF/none-10      59.11n ± 23%   60.27n ± 4%       ~ (p=0.485 n=6)
DeserializeNative/STEF/none-10    26.38n ±  1%   26.27n ± 1%       ~ (p=0.818 n=6)
SerializeFromPdata/STEF/none-10   1.349µ ±  1%   1.363µ ± 4%       ~ (p=0.104 n=6)
DeserializeToPdata/STEF/none-10   339.0n ± 17%   338.3n ± 2%       ~ (p=0.615 n=6)
geomean                           163.4n         164.4n       +0.59%

                                │     main     │                new                 │
                                │     B/op     │     B/op      vs base              │
SerializeNative/STEF/none-10      3.527Mi ± 0%   3.528Mi ± 0%       ~ (p=1.000 n=6)
DeserializeNative/STEF/none-10    836.2Ki ± 0%   836.2Ki ± 0%       ~ (p=0.900 n=6)
SerializeFromPdata/STEF/none-10   124.6Mi ± 0%   124.6Mi ± 0%       ~ (p=0.485 n=6)
DeserializeToPdata/STEF/none-10   29.81Mi ± 0%   29.81Mi ± 0%       ~ (p=0.563 n=6)
geomean                           10.17Mi        10.17Mi       +0.01%

                                │    main     │                 new                 │
                                │  allocs/op  │  allocs/op   vs base                │
SerializeNative/STEF/none-10      2.777k ± 0%   2.777k ± 0%       ~ (p=0.913 n=6)
DeserializeNative/STEF/none-10    1.230k ± 0%   1.230k ± 0%       ~ (p=1.000 n=6) ¹
SerializeFromPdata/STEF/none-10   256.3k ± 0%   256.3k ± 0%       ~ (p=0.818 n=6)
DeserializeToPdata/STEF/none-10   623.3k ± 0%   623.3k ± 0%       ~ (p=1.000 n=6) ¹
geomean                           27.18k        27.18k       +0.00%
```